requirements/testing.txt
coverage==4.0.3
Django==1.9.1
selenium==2.48.0
wheel==0.24.0
psycopg2==2.6.1
Pillow==3.0.0
djangorestframework==3.3.2
